modify ServicesModule
