+++ AElf.Benchmark/Program.cs
@@ -111,7 +111,7 @@ namespace AElf.Benchmark
                 builder.RegisterModule(new DatabaseAutofacModule());
                 builder.RegisterModule(new LoggerModule());
                 builder.RegisterModule(new StorageModule());
-                builder.RegisterModule(new ServicesModule());
+                builder.RegisterModule(new ChainAutofacModule());
                 builder.RegisterModule(new KernelAutofacModule());
                 builder.RegisterModule(new SmartContractAutofacModule());
                 builder.RegisterType<Benchmarks>().WithParameter("options", opts);
