+++ AElf.Kernel.Tests/ConfigureTestFramework.cs
@@ -47,7 +47,7 @@ namespace AElf.Kernel.Tests
             builder.RegisterModule(new SmartContractAutofacModule());
             builder.RegisterModule(new StateDictatorModule());
             builder.RegisterModule(new StorageModule());
-            builder.RegisterModule(new ServicesModule());
+            builder.RegisterModule(new ChainAutofacModule());
             builder.RegisterModule(new KernelAutofacModule());
             builder.RegisterInstance(new TxPoolConfig()).As<ITxPoolConfig>();
             builder.RegisterType<ContractTxPool>().As<IContractTxPool>().SingleInstance();
