+++ AElf.Miner.Tests/ConfigureTestFramework.cs
@@ -49,8 +49,9 @@ namespace AElf.Miner.Tests
             builder.RegisterModule(new SmartContractAutofacModule());
             builder.RegisterModule(new StateDictatorModule());
             builder.RegisterModule(new StorageModule());
-            builder.RegisterModule(new ServicesModule());
+            builder.RegisterModule(new ChainAutofacModule());
             builder.RegisterModule(new KernelAutofacModule());
+            builder.RegisterModule(new SmartContractAutofacModule());
             builder.RegisterInstance(new TxPoolConfig()).As<ITxPoolConfig>();
             builder.RegisterType<ContractTxPool>().As<IContractTxPool>().SingleInstance();
             builder.RegisterType<TxPoolService>().As<ITxPoolService>().SingleInstance();
